src/metrics/classification_metrics.py
from __future__ import annotations

import numpy as np
from sklearn.metrics import accuracy_score, f1_score


def compute_classification_metrics(y_true: np.ndarray, y_pred: np.ndarray) -> dict:
    return {
        'accuracy': float(accuracy_score(y_true, y_pred)),
        'macro_f1': float(f1_score(y_true, y_pred, average='macro')),
    }
