requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
statsmodels>=0.14
matplotlib>=3.8
tqdm>=4.66
torch>=2.2
torchvision>=0.17
pyyaml>=6.0
jupyter>=1.0
ipykernel>=6.29
